Split index.tsx util file into style.tsx and accessibility.ts
