+++ packages/components/src/components/Alert/Alert.tsx
@@ -13,7 +13,7 @@ import { BaseColor, useColorScheme } from '../../utils'
 import { Button, ButtonProps, ButtonVariants } from '../Button/Button'
 import { Icon, IconProps } from '../Icon/Icon'
 import { Spacer } from '../Spacer/Spacer'
-import { useTheme } from '../../utils/useTheme'
+import { useTheme } from '../../utils'
 
 /** Convenience function to set children content color correctly with light/dark mode */
 export const AlertContentColor = BaseColor
