+++ packages/components/src/utils/index.ts
@@ -1,0 +1,3 @@
+export * from './accessibility'
+export * from './OSfunctions'
+export * from './style'
